Memory mapping!
Before entering the C code we now produce a memory map of the system memory. The code used for this was taken from wiki.osdev.org. The code needs to be formatted to follow the rest of the respository. It works just fine on hardware, having had multiple successful trials. Memory maps for QEMU/BOCHS/* have been produced in addition with a memory map of my test system.