Kernel beginning
Replaced the old Kernel.ld file with a new, simpler version.

Modified the kernel loading process to use the correct cylinder values, I had forgotten to shift the cylinder bits across as the cylinder value is 10bits long, this was leading to the wrong cylinder to be read.

Added kernel entries to the makefile

Modified the kernel outline with further items such as the potential of VGA graphics, to be followed by VESA and the IDT. -- The IDT will be the next thing to be added to the kernel as it is required for elements such as system calls.