Keyboard driver
Implemented a basic keyboard driver which is capable of handling all
regular keys following the US QWERTY (the scancodes are the same for GB
keyboards in my case) 0-9, a-z and basic symbols.

The driver still needs the ability for control keys to be implemented in
addition for a method of getting a key when it's pressed on demand
(reference getch).

The keyboard layout can be configured using a header file in
RedFox32/Source/Kernel/KeyboardLayouts/<LayoutName>.h ensuring that all
header guards and variables are named exactly the same. Matching header
guards ensures that only the first layout is used and naming ensures no
code will break.

Old source has been deleted as it is no longer required for referencing.