Smaller code potential
It may be possible to enhance performance of the clear screen and scroll screen code using the rep instruction prefix, however it appears that in most cases, the standard implementations of memcpy and memset will be faster. This needs further investigation and potential identification of changes between old and new sources/hardware.