Keyboard driver events.
Added an "event" system for the keyboard driver. The Keyboard_GetEvent
method can be called to retrieve a `struct KeyboardEvent` which will
contain all information about the pressed key. Modifier keys do not get
added to the event but rather modify a `struct KeyboardModifiers` these
modifiers are entered into the event structure as `.Modifiers` using the
previously mentioned struct. These structs can be found in
`Source/Kernel/Keyboard.h` as they are required in places other than
`Source/Kernel/Keyboard.c`.

The event system allows for an interrupt driven keyboard to be used
however only grab the input when polling the event. Using this system we
are limited to a single keypress at a time (Excluding modifiers which
can be pressed in conjunction with other characters). Ultimately this
"limitation" will not affect generic use, however could be something to
look at in the future in cases such as if a user wanted to create a
game. This is easily solved by some form of input queue processing
however that is not a concern for the time being.

The driver now also changes the keyboard lights to indicate the three
lock keys. Caps lock toggles the capital mode and is XOR'd with shifts,
this creates a behavious where if both caps lock is enabled and a shift
key is down then the result will be lower case. This is non-standard
behavious however it is a behavious that I think can be useful. Num
lock and scroll lock do not currently perform any modification as the
test hardware I am using uses a built in keyboard which does not feature
this extended keys in a regular way. It uses the Fn key and a
combination of letter keys.