Keyboard performance
It came to my attention that the keyboard driver performance was
shockingly bad. I went through and reduced the amount of data being
duplicated in addition to adding a "Ready" value to the Event (this
could probably removed from the event however with future planning it
may be useful to maintain this value.