Allocation - Free - Safety
Added a 32 bit header to the `MemoryAllocation` structure:
`MemoryAllocation.AllocationHeader = 0xA110C000`.

Added the ability to `free` memory using `MemoryAllocation.Flags`; Bit 1
will be set for in use or clear for free. Upon freeing memory, the
allocation header ID is incremented preventing future frees but also
adding a second layer or protection to the memory.