Stack position
Moved the stack to exist in the region `0x00100000-0x00105000` (20KB).
This helps to reduce issues when constant data stored in the lower
sections of memory. It may also be worth moving the kernel up into this
region and shifting the stack further up into memory however this is for
future consideration.